+++ luigi/contrib/notifiers/__init__.py
@@ -1,0 +1,1 @@
+from luigi.contrib.notifiers.msteams import MsTeamsNotification, NotifiedTaskMixin
